+++ src/ColumnResizer.js
@@ -1,13 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import cn from 'classnames';
-import { DraggableCore } from 'react-draggable';
 
-import { noop, throttle } from './utils';
+import { noop } from './utils';
 
 const INVALID_VALUE = null;
-const MIN_WIDTH = 25;
-const THROTTLE_WAIT = 50;
 
 /**
  * ColumnResizer for BaseTable
@@ -16,58 +12,98 @@ class ColumnResizer extends React.PureComponent {
   constructor(props) {
     super(props);
 
+    this.isDragging = false;
     this.lastX = INVALID_VALUE;
     this.width = 0;
 
-    this._handleDrag = throttle(this._handleDrag.bind(this), THROTTLE_WAIT);
-    this._handleStart = this._handleStart.bind(this);
-    this._handleStop = this._handleStop.bind(this);
+    this._setHandleRef = this._setHandleRef.bind(this);
     this._handleClick = this._handleClick.bind(this);
+    this._handleMouseDown = this._handleMouseDown.bind(this);
+    this._handleMouseUp = this._handleMouseUp.bind(this);
+    this._handleMouseMove = this._handleMouseMove.bind(this);
+  }
+
+  componentWillMount() {
+    if (this.handleRef) {
+      const { ownerDocument } = this.handleRef;
+      ownerDocument.removeEventListener('mousemove', this._handleMouseMove);
+      ownerDocument.addEventListener('mouseup', this._handleMouseUp);
+    }
   }
 
   render() {
-    const { className, style, width, disabled } = this.props;
+    const { style, column, onResizeStart, onResize, onResizeStop, ...rest } = this.props;
 
-    const cls = cn('BaseTable__column-resizer', className);
     return (
-      <DraggableCore
-        axis="x"
-        disabled={disabled}
-        onStart={this._handleStart}
-        onDrag={this._handleDrag}
-        onStop={this._handleStop}
-      >
-        <div
-          className={cls}
-          onClick={this._handleClick}
-          style={{
-            position: 'absolute',
-            top: 0,
-            bottom: 0,
-            right: 0,
-            cursor: 'ew-resize',
-            borderRightStyle: 'solid',
-            ...style,
-            borderRightWidth: width,
-          }}
-        />
-      </DraggableCore>
+      <div
+        {...rest}
+        ref={this._setHandleRef}
+        onClick={this._handleClick}
+        onMouseDown={this._handleMouseDown}
+        onMouseUp={this._handleMouseUp}
+        style={{
+          userSelect: 'none',
+          touchAction: 'none',
+          position: 'absolute',
+          top: 0,
+          bottom: 0,
+          right: 0,
+          cursor: 'col-resize',
+          ...style,
+        }}
+      />
     );
   }
 
-  _handleDrag(e, data) {
+  _setHandleRef(ref) {
+    this.handleRef = ref;
+  }
+
+  _handleClick(e) {
+    e.stopPropagation();
+  }
+
+  _handleMouseDown(e) {
+    if (typeof e.button === 'number' && e.button !== 0) return;
+
+    this.isDragging = true;
+    this.lastX = INVALID_VALUE;
+    this.width = this.props.column.width;
+    this.props.onResizeStart(this.props.column);
+
+    const { ownerDocument } = this.handleRef;
+    ownerDocument.addEventListener('mousemove', this._handleMouseMove);
+    ownerDocument.addEventListener('mouseup', this._handleMouseUp);
+  }
+
+  _handleMouseUp(e) {
+    if (!this.isDragging) return;
+    this.isDragging = false;
+
+    this.props.onResizeStop(this.props.column);
+
+    const { ownerDocument } = this.handleRef;
+    ownerDocument.removeEventListener('mousemove', this._handleMouseMove);
+    ownerDocument.addEventListener('mouseup', this._handleMouseUp);
+  }
+
+  _handleMouseMove(e) {
+    const { offsetParent } = this.handleRef;
+    const offsetParentRect = offsetParent.getBoundingClientRect();
+    const x = e.clientX + offsetParent.scrollLeft - offsetParentRect.left;
+
     if (this.lastX === INVALID_VALUE) {
-      this.lastX = data.lastX;
-      return true;
+      this.lastX = x;
+      return;
     }
 
-    const { column } = this.props;
+    const { column, minWidth: MIN_WIDTH } = this.props;
     const { width, maxWidth, minWidth = MIN_WIDTH } = column;
-    const movedX = data.x - this.lastX;
-    if (!movedX) return true;
+    const movedX = x - this.lastX;
+    if (!movedX) return;
 
     this.width = this.width + movedX;
-    this.lastX = data.x;
+    this.lastX = x;
 
     let newWidth = this.width;
     if (maxWidth && newWidth > maxWidth) {
@@ -76,53 +112,27 @@ class ColumnResizer extends React.PureComponent {
       newWidth = minWidth;
     }
 
-    if (newWidth === width) return true;
-    return this.props.onResize(column, newWidth);
-  }
-
-  _handleStart(e, data) {
-    this.lastX = INVALID_VALUE;
-    this.width = this.props.column.width;
-    return this.props.onResizeStart(this.props.column);
-  }
-
-  _handleStop(e, data) {
-    return this.props.onResizeStop(this.props.column);
-  }
-
-  _handleClick(e) {
-    e.stopPropagation();
+    if (newWidth === width) return;
+    this.props.onResize(column, newWidth);
   }
 }
 
 ColumnResizer.defaultProps = {
-  width: 3,
   onResizeStart: noop,
   onResize: noop,
   onResizeStop: noop,
+  minWidth: 30,
 };
 
 ColumnResizer.propTypes = {
   /**
-   * Class name for the drag handler
+   * Custom style for the drag handler
    */
-  className: PropTypes.string,
+  style: PropTypes.object,
   /**
    * The column object to be dragged
    */
   column: PropTypes.object,
-  /**
-   * The width of the drag handler
-   */
-  width: PropTypes.number.isRequired,
-  /**
-   * Whether the resizing is disabled
-   */
-  disabled: PropTypes.bool,
-  /**
-   * Custom style for the drag handler
-   */
-  style: PropTypes.object,
   /**
    * A callback function when resizing started
    * The callback is of the shape of `(column) => *`
@@ -138,6 +148,10 @@ ColumnResizer.propTypes = {
    * The callback is of the shape of `(column) => *`
    */
   onResizeStop: PropTypes.func,
+  /**
+   * Minimum width of the column could be resized to if the column's `minWidth` is not set
+   */
+  minWidth: PropTypes.number,
 };
 
 export default ColumnResizer;